requirements.txt
psycopg2-binary>=2.9.0
snowflake-connector-python>=3.0.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0